+++ NEWS.md
@@ -1,11 +1,20 @@
+# wildcam 0.4.0.9000 (2019-02-21)
+
+* Add `eval_pt` and `eval_buffer` for evaluating and characterizing camera trap locations [!4](https://gitlab.com/robit.a/wildcam/merge_requests/4).
+* Add "wetland" and "lc" example data. 
+
 # wildcam 0.3.0.9000 (2019-02-15)
 
 * Add `make_grid` for making camera trap grids around focal cameras [!3](https://gitlab.com/robit.a/wildcam/merge_requests/3).
 
+
 # wildcam 0.2.0.9000 (2019-02-12)
 
 * Add `dist_to` for distance to calculations [!2](https://gitlab.com/robit.a/wildcam/merge_requests/2). 
+* Add "water" example data.
+
 
 # wildcam 0.1.0.9000 (2019-02-11)
 
 * Add `strat_sample` for stratified polygon sampling [!1](https://gitlab.com/robit.a/wildcam/merge_requests/1). 
+* Add "points" and "densitygrid" example data.